greeter-handler/requirements.txt
apache-flink-statefun==3.0.0